# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "QRFLOW.codes API",
        "version": "1.0.0",
        "summary": "Create QR codes, change where dynamic codes point, pull scan analytics, bulk-generate, manage webhooks.",
        "description": "Everything the QRFLOW.codes dashboard does, from your own code. Authenticate with a Business API key (`Authorization: Bearer qrf_live_\u2026`, made on the Account page) or an OAuth token from https://qrflow.codes/oauth/authorize. Dynamic codes (url, phone, email, sms, location on paid plans) can be re-pointed after printing; their printed link uses the account's own domain when one is connected. Also available as an MCP server at https://qrflow.codes/mcp for Claude, ChatGPT, Cursor and Claude Code.",
        "termsOfService": "https://qrflow.codes/terms",
        "contact": {
            "name": "QRFLOW.codes",
            "url": "https://qrflow.codes/developers",
            "email": "hello@qrflow.codes"
        }
    },
    "servers": [
        {
            "url": "https://qrflow.codes/api/v1"
        }
    ],
...
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
...
        "/codes": {
            "get": {
                "tags": [
                    "codes"
                ],
                "summary": "List codes, newest first",
                "operationId": "listCodes",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "maximum": 100,
                            "default": 50
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search labels",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "codes": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Code"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Missing or invalid key",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Key lacks codes:read",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "codes"
                ],
                "summary": "Create a code",
                "operationId": "createCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCode"
                            },
                            "examples": {
                                "url": {
                                    "value": {
                                        "type": "url",
                                        "destination_data": {
                                            "url": "https://example.com/menu"
                                        },
                                        "label": "Table tents",
                                        "frame_style": "caption-below",
                                        "frame_caption": "Scan for menu"
                                    }
                                },
                                "wifi": {
                                    "value": {
                                        "type": "wifi",
                                        "destination_data": {
                                            "ssid": "CafeGuest",
                                            "password": "latte-2026",
                                            "encryption": "WPA"
                                        },
                                        "label": "Counter card"
                                    }
                                },
                                "googleReview": {
                                    "value": {
                                        "type": "url",
                                        "destination_data": {
                                            "subtype": "googlereview",
                                            "placeId": "ChIJN1t_tDeuEmsRUsoyG83frY4"
                                        },
                                        "label": "Receipt footer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "code": {
                                            "$ref": "#/components/schemas/Code"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid fields or destination, or the saved-codes ceiling",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "402": {
                        "description": "The plan does not include this kind",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "A Business API key (qrf_live_\u2026) or an OAuth access token."
            },
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "https://qrflow.codes/oauth/authorize",
                        "tokenUrl": "https://qrflow.codes/api/oauth/token",
                        "refreshUrl": "https://qrflow.codes/api/oauth/token",
                        "scopes": {
                            "profile": "Plan and email",
                            "codes:read": "Read codes",
                            "codes:write": "Create, edit, delete codes",
                            "analytics:read": "Scan analytics",
                            "domains:read": "Link domains",
                            "webhooks:manage": "Webhooks"
                        }
                    }
                }
            }
        },
        "schemas": {
            "Error": {
                "type": "object",
                "properties": {
                    "error": {
                        "type": "string",
                        "examples": [
                            "invalid_request",
                            "invalid_token",
                            "upgrade_required",
                            "insufficient_scope",
                            "not_found",
                            "conflict",
                            "rate_limited"
                        ]
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
...
            "CreateCode": {
                "type": "object",
                "required": [
                    "type",
                    "destination_data"
                ],
                "properties": {
                    "type": {
                        "type": "string",
                        "enum": [
                            "url",
                            "text",
                            "wifi",
                            "vcard",
                            "email",
                            "phone",
                            "sms",
                            "location"
                        ]
                    },
                    "destination_data": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "description": "The fields for the type; add `subtype` (from /catalog) for a specific kind such as instagram or googlereview"
                    },
                    "label": {
                        "type": "string",
                        "maxLength": 120
                    },
                    "fg_color": {
                        "type": "string",
                        "pattern": "^#[0-9a-fA-F]{6}$",
                        "example": "#0F172A"
                    },
                    "bg_color": {
                        "type": "string",
                        "pattern": "^#[0-9a-fA-F]{6}$",
                        "example": "#0F172A"
                    },
                    "frame_style": {
                        "type": "string"
                    },
                    "frame_caption": {
                        "type": "string",
                        "maxLength": 60
                    },
                    "frame_caption2": {
                        "type": "string",
                        "maxLength": 60
                    },
                    "domain_id": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "uuid"
                    }
                }
            },
...
            "Code": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "label": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "kind": {
                        "type": "string"
                    },
                    "kind_label": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "destination_data": {
                        "type": "object"
                    },
                    "destination": {
                        "type": "string"
                    },
                    "dynamic": {
                        "type": "boolean"
                    },
                    "dynamic_capable": {
                        "type": "boolean"
                    },
                    "short_code": {
                        "type": "string"
                    },
                    "short_url": {
                        "type": "string",
                        "description": "What to print for a dynamic code"
                    },
                    "domain_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "fg_color": {
                        "type": "string"
                    },
                    "bg_color": {
                        "type": "string"
                    },
                    "has_logo": {
                        "type": "boolean"
                    },
                    "frame_style": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "frame_caption": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "scans": {
                        "type": "integer"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "manage_url": {
                        "type": "string"
                    },
                    "image_url": {
                        "type": "string"
                    }
                }
            },
...
        }
    },
...
}